Create faherheit_to_celsius.c
This program uses a method more secure against buffer overflows to convert between faherheit adn celsisus, I was working the first chapter of K&R but knew if I wanted to impress prospective employers by creating something I was going to have to knock it out of the park and do something the boot didn't tell me to do right off the bat so  I found a blog with a method  to do interget input  in a  fashion apparentally concenred with preventing buffer overflow explots in c. Of course the floats-input.h was about changing his input to work with floats and was actually done by Open AI's chat GPT, because I'm  practing with K&R, but I thought doing it this way when i'm brand new might impress prospective employers.   It takes input from the user, and converts Faherheit to Celsius.

The united states doesn't use the metric system, we should. But Eh...  One problem down as far as that's concerned, you'll never have to do the conversion by handagain, you can use a machine to approximate. I've found Google's doesn't suffer  the Rounding error mine does, but  C does a close enough job to be useful to most people, and i'm sure an experiencced developer could work mine to make not round off as much.